% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l01.R
\name{biasofsd}
\alias{biasofsd}
\title{Calculate the bias incurred by when using the sample standard deviation to 
estimate the population value.}
\usage{
biasofsd()
}
\description{
Calculate the bias incurred by when using the sample standard deviation to 
estimate the population value.
}
