% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialciplot.R
\name{binomialciplot}
\alias{binomialciplot}
\title{Construct confidence intervals for binomial probability.}
\usage{
binomialciplot(n = 7, alpha = 0.05, npts = 10000)
}
\arguments{
\item{n}{number of independent observations}

\item{alpha}{Error for confidence interval; interval has level 1-alpha.}

\item{npts}{number of grid points for cdf inversion.}
}
\description{
Construct confidence intervals for binomial probability.
}
\examples{
binomialciplot(alpha=0.2)
}
