% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binomialglrt.R
\name{binomialglrt}
\alias{binomialglrt}
\title{Display the generalized likelihood ratio statistic for a simple null hypothesis about a binomial distribution.}
\usage{
binomialglrt()
}
\description{
Display the generalized likelihood ratio statistic for a simple null hypothesis about a binomial distribution.
}
\examples{
binomialglrt()
}
