% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l07.R
\name{binpowcurve}
\alias{binpowcurve}
\title{Plot power for an exact binomial test.}
\usage{
binpowcurve(m, pi0, alpha)
}
\arguments{
\item{m}{sample size for binomial test.}

\item{pi0}{null hypothesis value}

\item{alpha}{test level.}
}
\description{
Plot power for an exact binomial test.
}
\examples{
binpowcurve(10,.5,.05)
}
