% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariateabincover.R
\name{bivariatebincover}
\alias{bivariatebincover}
\title{Plot coverage of various confidence intervals for the difference between binomial proportions.}
\usage{
bivariatebincover(n1 = 10, n2 = 10, npts = 99)
}
\arguments{
\item{n1}{size of first binomial sample}

\item{n2}{size of second binomial sample}

\item{npts}{number of grid points on each side of the pi1xpi2 grid}
}
\description{
Plot coverage of various confidence intervals for the difference between binomial proportions.
}
