% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccplot.R
\name{ccplot}
\alias{ccplot}
\title{Draw a plot exhibiting the continuity correction}
\usage{
ccplot(nn = 10, x = 4)
}
\arguments{
\item{nn}{Binomial sample size}

\item{x}{number of successes}
}
\description{
Draw a plot exhibiting the continuity correction
}
