% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l03.R
\name{checkdist}
\alias{checkdist}
\title{Using simulation, check the operating characteristics of two sample 
pooled and unpooled t tests, under various assumptions on the
ratio of group standard deviations.}
\usage{
checkdist(m, n, nsamp = 10000, nrat = 5)
}
\arguments{
\item{m}{number of independent observations in first group.}

\item{n}{number of independent observations in second group.}

\item{nsamp}{number of samples.}

\item{nrat}{number of standard deviation ratios to explore, beginning at 1, ending at 10}
}
\value{
A list with components
out a matrix with nrat rows and two columns giving counts of samples with
      type I errors.
rat the vector of sd ratios
nsamp copy of the input nsamp
m copy of the input m
n copy of the input n
}
\description{
Using simulation, check the operating characteristics of two sample 
pooled and unpooled t tests, under various assumptions on the
ratio of group standard deviations.
}
\examples{
a<-checkdist(4,4,nsamp=1000)
}
