% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corllkp.R
\name{corllkp}
\alias{corllkp}
\title{Plot log likelihood for normal correlation, known expectations and variances.}
\usage{
corllkp(Sxx, Sxy, Syy, n)
}
\arguments{
\item{Sxx}{average of squares of first variable}

\item{Sxy}{average of product of first and second variables}

\item{Syy}{average of squares of second variable}

\item{n}{sample size.}
}
\description{
Plot log likelihood for normal correlation, known expectations and variances.
}
\examples{
corllkp(1.5,.5,1.5,1)
}
