% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctsci.R
\name{ctsci}
\alias{ctsci}
\title{Draw graphical construction for confidence interval for continuous distribution.}
\usage{
ctsci(dist = qbeta, alpha = 0.05)
}
\arguments{
\item{dist}{quantile function for construction of confidence interval.}

\item{alpha}{1-coverage for confidence interval.}
}
\description{
Draw graphical construction for confidence interval for continuous distribution.
}
\examples{
ctsci()
}
