% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allcover.R
\name{fun.allcover}
\alias{fun.allcover}
\title{Calculate true coverage probabilities for a variety of nominal binomial confidence intervals}
\usage{
fun.allcover(nn, larange = c(-6, -1), npts = 50000, plotdat = NULL)
}
\arguments{
\item{nn}{number of binomial samples.}

\item{npts}{number of potential nominal coverages to consider}

\item{plotdat}{optional saved output for replotting.}

\item{larage}{range for log of nominal coverage}
}
\description{
Calculate true coverage probabilities for a variety of nominal binomial confidence intervals
}
