% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{fun.allwald}
\alias{fun.allwald}
\title{All Wald tests for the differencce in two binomial coefficients.}
\usage{
fun.allwald(n1, n2)
}
\arguments{
\item{n1}{Number of trials from group 1}

\item{n2}{Number of trials from group 1}
}
\description{
All Wald tests for the differencce in two binomial coefficients.
}
