% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approxnormalci.R
\name{fun.approxnormalci}
\alias{fun.approxnormalci}
\title{Give basic normal approximation to the binomial confidence interval}
\usage{
fun.approxnormalci(x, nn, alpha = 0.05, alternative = "two.sided")
}
\arguments{
\item{x}{number of succsses}

\item{nn}{number of trials}

\item{alpha}{1-coverage probability}

\item{alternative}{character string indicating alternative}
}
\description{
Give basic normal approximation to the binomial confidence interval
}
