% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverageplot.R
\name{fun.coverageplot}
\alias{fun.coverageplot}
\title{Calculate, and plot, coverage for binomial confidence intervals.}
\usage{
fun.coverageplot(
  nn,
  alpha = 0.05,
  exactonly = FALSE,
  alternative = "two.sided",
  plot = TRUE,
  cm = 0,
  npi = 999
)
}
\arguments{
\item{nn}{number of binomial samples}

\item{alpha}{1-coverage}

\item{alternative}{string incidating whether one- or two sided intervals are calculated}

\item{plot}{logical flag indicating whether to plot.}

\item{cm}{minimal value on vertical axis.}

\item{npi}{Number of pi vectors to try}
}
\description{
Calculate, and plot, coverage for binomial confidence intervals.
}
