% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactbin.R
\name{fun.exactbin}
\alias{fun.exactbin}
\title{Exact binoial confidence intervals via the F distribution}
\usage{
fun.exactbin(x0, x1, level = 0.95)
}
\arguments{
\item{x0}{Number of successes}

\item{x1}{Number of failures}

\item{level}{Interval level}
}
\description{
Exact binoial confidence intervals via the F distribution
}
