% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottestpoly.R
\name{fun.plottestpoly}
\alias{fun.plottestpoly}
\title{Plot the results of simulating polychoric correlation and kappa behavior.}
\usage{
fun.plottestpoly(simout)
}
\arguments{
\item{simout}{Output from fun.testpoly}
}
\description{
Plot the results of simulating polychoric correlation and kappa behavior.
}
