% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratsize.R
\name{fun.stratsize}
\alias{fun.stratsize}
\title{Simulate incorrect two-stage testing for effect between
exposure and result, by pretesting for the effect of 
a confounder, and adjusting if statistical evidence
says to.}
\usage{
fun.stratsize(qv, cut, r1, h)
}
\arguments{
\item{qv}{vector of cell probabilities, going down column}

\item{cut}{Critical value.}

\item{r1}{row 1 total}

\item{h}{table total.}
}
\description{
Simulate incorrect two-stage testing for effect between
exposure and result, by pretesting for the effect of 
a confounder, and adjusting if statistical evidence
says to.
}
