% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testncp.R
\name{fun.testncp}
\alias{fun.testncp}
\title{Test the noncentral chi-square approximation to Pearson's statistic}
\usage{
fun.testncp(p0, p1, size = 10, nsamp = 1e+05)
}
\arguments{
\item{p0}{cell probabilities under null}

\item{p1}{cell probabilities under alternative}

\item{size}{number of items classified}

\item{nsamp}{number of Monte Carlo samples.}
}
\description{
Test the noncentral chi-square approximation to Pearson's statistic
}
