% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testpoly.R
\name{fun.testpoly}
\alias{fun.testpoly}
\title{Simulate data from a bivariate normal distribution, and 
calculate various measures of association.}
\usage{
fun.testpoly(rho = 0.5, nsamp = 50, nmc = 10000, progress = F)
}
\arguments{
\item{rho}{Correlation of marginal normals}

\item{nsamp}{Number of observations in each sample}

\item{nmc}{Number of Monte Carlo simulations}

\item{progress}{Logical indicating whether periodic updates of progress should be made.}
}
\value{
list with two components: mat, a matrix with seven columns representing polychoric correlation with 2, 3, and 4 groupings, Pearson correlation, and kappa statistic for 2, 3, and 4 categories., and settings, rho and nsamp.
}
\description{
Simulate data from a bivariate normal distribution, and 
calculate various measures of association.
}
