% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{fun.wald}
\alias{fun.wald}
\title{Wald test for the differencce in two binomial coefficients.}
\usage{
fun.wald(y1, y2, n1, n2)
}
\arguments{
\item{y1}{Number of successes from group 1}

\item{y2}{Number of successes from group 1}

\item{n1}{Number of trials from group 1}

\item{n2}{Number of trials from group 1}
}
\description{
Wald test for the differencce in two binomial coefficients.
}
