% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l10.R
\name{inference}
\alias{inference}
\title{Perform inference on conditional sample space.}
\usage{
inference(out, alpha = 0.05, rng = c(-5, 5))
}
\arguments{
\item{out}{List of the sort provided by network.}

\item{alpha}{Test level, or 1- confidence level.}

\item{rng}{Range of possible parameter values.}
}
\value{
List with components:
  ospv: Observed one-sided p values
  tspv: Observed two-sided p value.
  ci: confidence interval.
  mle: Maximum likelihood estimator.
}
\description{
Examples apply this to data from \insertCite{mehtapatel;textual}{TheorStat}, citing \insertCite{goorinetal87;textual}{TheorStat}.
}
\examples{
#Columns in table are:
# Lymphocytic Infiltration (1=low, 0=high)
# Sex (1=male, 0=female)
# Any Ostioid Pathology (1=yes, 0=no)
# Number in LI-Sex-AOP group
# Number in LI-Sex-AOP group with disease free interval greater than 3 y
goorin<-data.frame(LI=c(0,0,0,0,1,1,1,1),Sex=c(0,0,1,1,0,0,1,1),
   AOP=c(0,1,0,1,0,1,0,1),N=c(3,2,4,1,5,5,9,17),Y=c(3,2,4,1,5,3,5,6))
print(goorin)
\donttest{
out<-network(goorin[,1:3],goorin[,4],conditionon=1:3,resp=goorin[,5])
inference(out)
}
}
\references{
\insertRef{mehtapatel}{TheorStat}

\insertRef{goorinetal87}{TheorStat}
}
