% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l05.R
\name{llkmaximization}
\alias{llkmaximization}
\title{Function exhibiting Newton method for maximizing an exponential log likelihood.}
\usage{
llkmaximization(X = 1.5, lambda0 = 1)
}
\arguments{
\item{X}{exponential observation.}

\item{lambda0}{Initial exponential parameter.}
}
\description{
Function exhibiting Newton method for maximizing an exponential log likelihood.
}
\examples{
llkmaximization()
}
