% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanmedianpower.R
\name{meanmedianpower}
\alias{meanmedianpower}
\title{Calculate power for exact mean and median tests for normal variates.  
Critical values are cacluated via Monte Carlo.}
\usage{
meanmedianpower(n = 9, nsamp = 5000, nalt = 100)
}
\arguments{
\item{n}{number of observations in sample}

\item{nsamp}{number of MC samples}

\item{nalt}{number of alternatives to consider, greater than 0 and less
than or equal to 1.}
}
\description{
Calculate power for exact mean and median tests for normal variates.  
Critical values are cacluated via Monte Carlo.
}
\examples{
meanmedianpower()
}
