% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkdist.R
\name{plotcheckdist}
\alias{plotcheckdist}
\title{Plot the results of the function checkdist, which determined the true level of
pooled and unpooled t tests.}
\usage{
plotcheckdist(checkdistout)
}
\arguments{
\item{checkdistout}{a list with components
out a matrix with nrat rows and two columns giving counts of samples with
      type I errors.
rat the vector of sd ratios
nsamp number of samples out of which the matrix out was recorded.
m number in first group.
n number in second group.}
}
\description{
Plot the results of the function checkdist, which determined the true level of
pooled and unpooled t tests.
}
\examples{
plotcheckdist(checkdist(4,4,nsamp=2000,nrat=5))
}
