% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l05.R
\name{plotllkcontour}
\alias{plotllkcontour}
\title{Plot countours of the log likelihood for a logistic regression
with an intercept parameter and one linear parameter.  
Examples apply this to data from \insertCite{mehtapatel;textual}{TheorStat}, citing \insertCite{goorinetal87;textual}{TheorStat}.}
\usage{
plotllkcontour(dataset, xvl, yyl, nnl)
}
\arguments{
\item{dataset}{data set name}

\item{xvl}{string with name of covariate in data set.}

\item{yyl}{string with name of response variable representing counts of 
successes in data set.}

\item{nnl}{string with name of variable having counts.}
}
\description{
Plot countours of the log likelihood for a logistic regression
with an intercept parameter and one linear parameter.  
Examples apply this to data from \insertCite{mehtapatel;textual}{TheorStat}, citing \insertCite{goorinetal87;textual}{TheorStat}.
}
\examples{
#Columns in table are:
# Lymphocytic Infiltration (1=low, 0=high)
# Sex (1=male, 0=female)
# Any Ostioid Pathology (1=yes, 0=no)
# Number in LI-Sex-AOP group
# Number in LI-Sex-AOP group with disease free interval greater than 3 y
goorin<-data.frame(LI=c(0,0,0,0,1,1,1,1),Sex=c(0,0,1,1,0,0,1,1),
   AOP=c(0,1,0,1,0,1,0,1),N=c(3,2,4,1,5,5,9,17),Y=c(3,2,4,1,5,3,5,6))
print(goorin)
glm(cbind(Y,N-Y)~Sex,data=goorin,family="binomial")
glm(cbind(Y,N-Y)~LI,data=goorin,family="binomial")
plotllkcontour(goorin,"LI","Y","N")
plotllkcontour(goorin,"AOP","Y","N")
}
\references{
\insertRef{mehtapatel}{TheorStat}

\insertRef{goorinetal87}{TheorStat}
}
