% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l09.R
\name{pvalpic}
\alias{pvalpic}
\title{Diagram for an example of a Neyman-Pearson test for two simple alternatives.
This example illustrates null and alternative distributions on [0,1].  Null 
is uniform.  Alternative has density flat on [0,.5) and flat on (.5,1], and
then given small slope.}
\usage{
pvalpic(gamma, epsilon)
}
\arguments{
\item{gamma}{amount added to density for alternative on (.5,1], and subtracted
from density on [0,.5).}

\item{epsilon}{slope added to alternative hypothesis density}
}
\description{
Diagram for an example of a Neyman-Pearson test for two simple alternatives.
This example illustrates null and alternative distributions on [0,1].  Null 
is uniform.  Alternative has density flat on [0,.5) and flat on (.5,1], and
then given small slope.
}
\examples{
pvalpic(.5,.05)
}
