% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tllk.R
\name{tllk}
\alias{tllk}
\title{Display the log likelihood function for location parameter for sample of 5 independent Cauchy variates.}
\usage{
tllk()
}
\description{
Display the log likelihood function for location parameter for sample of 5 independent Cauchy variates.
}
