% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l02.R
\name{uniformciplot}
\alias{uniformciplot}
\title{Construct confidence intervals for upper bound of uniform variables.}
\usage{
uniformciplot(n = 10, thetamax = 10, alpha = 0.05)
}
\arguments{
\item{n}{number of independent observations}

\item{thetamax}{upper bound for horizontal axis.}

\item{alpha}{Error for confidence interval; interval has level 1-alpha.}
}
\description{
Construct confidence intervals for upper bound of uniform variables.
}
\examples{
uniformciplot()
}
