% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceglrt.R
\name{varianceglrt}
\alias{varianceglrt}
\title{Diagram the generalized likelihood ratio test for the variance.}
\usage{
varianceglrt()
}
\description{
Diagram the generalized likelihood ratio test for the variance.
}
\examples{
varianceglrt()
}
