% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{waldsize}
\alias{waldsize}
\title{Rejection probabilities for pairs of binomial outcomes, on a grid of common success probabilities}
\usage{
waldsize(n1, n2, npts = 100)
}
\arguments{
\item{n1}{Number of trials from group 1}

\item{n2}{Number of trials from group 1}

\item{npts}{number of grid points}
}
\description{
Rejection probabilities for pairs of binomial outcomes, on a grid of common success probabilities
}
