% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wilsonpicture.R
\name{wilsonpicture}
\alias{wilsonpicture}
\title{Graphically display construction of Wilson confidence intervals for
binomial variables.}
\usage{
wilsonpicture(nn = 10, alpha = 0.05, pinull = 0.6, npts = 100, piobs = 0.5)
}
\arguments{
\item{nn}{binomial sample size}

\item{alpha}{1-coverage}

\item{pinull}{Potential null hypothesis for parameter}

\item{npts}{Number of points in elipse}

\item{piobs}{Observed proportion value.}
}
\description{
Graphically display construction of Wilson confidence intervals for
binomial variables.
}
